+++ packages/client/modules/demo/ClientGraphQLServer.ts
@@ -5,9 +5,12 @@ import StrictEventEmitter from 'strict-event-emitter-types'
 import handleCompletedDemoStage from './handleCompletedDemoStage'
 import {
   DragReflectionDropTargetTypeEnum,
-  IDiscussPhase,
+  IDiscussPhase, IGoogleAnalyzedEntity,
   INewMeetingStage,
   IReflectPhase,
+  IRetroReflection,
+  IRetroReflectionGroup,
+  ITask,
   NewMeetingPhase,
   NewMeetingPhaseTypeEnum
 } from '../../types/graphql'
@@ -29,9 +32,23 @@ import initDB, {demoMeetingId, demoTeamId, demoViewerId, GitHubProjectKeyLookup,
 import LocalAtmosphere from './LocalAtmosphere'
 import ms from 'ms'
 import Reflection from 'parabol-server/database/types/Reflection'
-import DemoReflection from './types/DemoReflection'
-import DemoReflectionGroup from './types/DemoReflectionGroup'
-import DemoTask from './types/DemoTask'
+import extractTextFromDraftString from '../../utils/draftjs/extractTextFromDraftString'
+import entityLookup from './entityLookup'
+import getDemoEntities from './getDemoEntities'
+import stringSimilarity from 'string-similarity'
+
+export type DemoReflection = Omit<IRetroReflection, 'autoReflectionGroupId' | 'team'> & {
+  creatorId: string
+  reflectionId: string
+  isHumanTouched: boolean
+}
+
+export type DemoReflectionGroup = Omit<IRetroReflectionGroup, 'team' | 'reflections'> & {
+  reflectionGroupId: string
+  reflections: DemoReflection[]
+}
+
+export type DemoTask = Omit<ITask, 'agendaItem'>
 
 interface Payload {
   __typename: string
@@ -254,32 +271,60 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
       }
       return {createJiraIssue: data}
     },
-    CreateReflectionMutation: (
+    CreateReflectionMutation: async (
       {input: {content, retroPhaseItemId, sortOrder, id, groupId}},
-      userId
+      userId: string
     ) => {
       const now = new Date().toJSON()
       const reflectPhase = this.db.newMeeting.phases![1] as IReflectPhase
       const phaseItem = reflectPhase.reflectPrompts.find((prompt) => prompt.id === retroPhaseItemId)
       const reflectionGroupId = groupId || this.getTempId('refGroup')
       const reflectionId = id || this.getTempId('ref')
-      const reflection = new DemoReflection({
+      const plaintextContent = extractTextFromDraftString(content)
+      let entities = [] as IGoogleAnalyzedEntity[]
+      if (userId !== demoViewerId) {
+        entities = entityLookup[reflectionId].entities
+      } else if (plaintextContent && plaintextContent.length > 2) {
+        entities = await getDemoEntities(plaintextContent)
+      }
+
+      const reflection = {
+        __typename: 'RetroReflection',
         id: reflectionId,
-        createdAt: now as any,
+        isHumanTouched: false,
+        reflectionId,
+        createdAt: now,
         creatorId: userId,
         content,
-        entities: [],
+        plaintextContent,
+        dragContext: null,
+        editorIds: [],
+        isActive: true,
+        isEditing: null,
+        isViewerCreator: userId === demoViewerId,
+        entities,
         meetingId: demoMeetingId,
+        meeting: this.db.newMeeting,
         phaseItem,
         reflectionGroupId,
         retroPhaseItemId,
         sortOrder: 0,
-        updatedAt: now as any
-      })
+        updatedAt: now,
+        retroReflectionGroup: undefined as any
+      } as DemoReflection
 
-      const reflectionGroup = new DemoReflectionGroup({
+      const smartTitle = getGroupSmartTitle([reflection])
+
+      const reflectionGroup = {
+        __typename: 'RetroReflectionGroup',
         id: reflectionGroupId,
-        createdAt: now as any,
+        reflectionGroupId,
+        smartTitle,
+        title: smartTitle,
+        voteCount: 0,
+        viewerVoteCount: 0,
+        createdAt: now,
+        isActive: true,
         meetingId: demoMeetingId,
         meeting: this.db.newMeeting,
         phaseItem,
@@ -287,9 +332,10 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
         reflections: [reflection],
         sortOrder,
         tasks: [],
-        updatedAt: now as any,
+        titleIsUserDefined: false,
+        updatedAt: now,
         voterIds: []
-      })
+      } as DemoReflectionGroup
 
       reflection.retroReflectionGroup = reflectionGroup as any
       this.db.reflectionGroups.push(reflectionGroup)
@@ -365,10 +411,26 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
       }
       return {removeReflection: data}
     },
-    UpdateReflectionContentMutation: ({reflectionId, content}, userId) => {
+    UpdateReflectionContentMutation: async ({reflectionId, content}, userId) => {
       const reflection = this.db.reflections.find((reflection) => reflection.id === reflectionId)!
       reflection.content = content
       reflection.updatedAt = new Date().toJSON()
+      const plaintextContent = extractTextFromDraftString(content)
+      const isVeryDifferent =  stringSimilarity.compareTwoStrings(plaintextContent, reflection.plaintextContent) < 0.9
+      const entities = isVeryDifferent ? await getDemoEntities(plaintextContent) : reflection.entities
+      reflection.plaintextContent = plaintextContent
+      reflection.entities = entities
+
+      const reflectionsInGroup = this.db.reflections.filter(({reflectionGroupId}) => reflectionGroupId === reflection.reflectionGroupId)
+      const newTitle = getGroupSmartTitle(reflectionsInGroup)
+      const reflectionGroup = this.db.reflectionGroups.find((group) => group.id === reflection.reflectionGroupId)
+      if (reflectionGroup) {
+        const titleIsUserDefined = reflectionGroup.smartTitle !== reflectionGroup.title
+        reflectionGroup.smartTitle = newTitle
+        if (!titleIsUserDefined) {
+          reflectionGroup.title = newTitle
+        }
+      }
       const data = {
         error: null,
         meeting: this.db.newMeeting,
@@ -511,29 +573,24 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
       }
       return {setStageTimer: data}
     },
-    StartDraggingReflectionMutation: ({reflectionId, initialCoords, dragId}, userId) => {
-      let dragCoords = initialCoords
+    StartDraggingReflectionMutation: ({reflectionId, dragId}, userId) => {
       const reflection = this.db.reflections.find((reflection) => reflection.id === reflectionId)!
       if (userId !== demoViewerId) {
-        if (reflection.isHumanTouched || !(this.atmosphere as any).getMasonry) return
-        const {itemCache} = (this.atmosphere as any).getMasonry()
-        const cachedItem = itemCache[reflectionId]
-        const bbox = (cachedItem.el as HTMLDivElement).getBoundingClientRect()
-        dragCoords = {x: bbox.left, y: bbox.top}
+        if (reflection.isHumanTouched) return
       } else {
         reflection.isHumanTouched = true
       }
+      const user = this.db.users.find((user) => user.id === userId)
       const data = {
         error: null,
         teamId: demoTeamId,
         meetingId: demoMeetingId,
         reflectionId,
         reflection,
-        dragContext: {
-          id: dragId || this.getTempId('drag'),
+        remoteDrag: {
+          id: dragId,
           dragUserId: userId,
-          dragUser: this.db.users.find((user) => user.id === userId),
-          dragCoords
+          dragUserName: user ? user.preferredName : 'A Ghost'
         },
         __typename: 'StartDraggingReflectionPayload'
       }
@@ -631,6 +688,7 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
             updatedAt: now
           })
           reflectionGroup.reflections!.push(reflection as any)
+          reflectionGroup.reflections.sort((a, b) => (a.sortOrder < b.sortOrder ? 1 : -1))
           oldReflections.splice(
             oldReflections.findIndex((reflection) => reflection === reflection),
             1
@@ -674,6 +732,7 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
           failedDrop = true
         }
       }
+      const user = this.db.users.find((user) => user.id === userId)
       const data = {
         __typename: 'EndDraggingReflectionPayload',
         meetingId: demoMeetingId,
@@ -689,7 +748,11 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
         userId,
         dropTargetType: failedDrop ? null : dropTargetType,
         dropTargetId: failedDrop ? null : dropTargetId,
-        dragId
+        remoteDrag: {
+          id: dragId,
+          dragUserId: userId,
+          dragUserName: user ? user.preferredName : 'A Ghost'
+        }
       }
 
       if (userId !== demoViewerId) {
@@ -699,7 +762,7 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
     },
     UpdateDragLocationMutation: ({input}, userId) => {
       const {teamId, ...inputData} = input
-      const data = {...inputData, userId, __typename: 'UpdateDragLocationPayload'}
+      const data = {remoteDrag: inputData, userId, __typename: 'UpdateDragLocationPayload'}
       if (userId !== demoViewerId) {
         const {sourceId} = inputData
         const reflection = this.db.reflections.find((reflection) => reflection.id === sourceId)
@@ -974,20 +1037,19 @@ class ClientGraphQLServer extends (EventEmitter as GQLDemoEmitter) {
           newMeetingMember.tasks.push(task as any)
         }
       }
-      const nextTask = new DemoTask({
-        ...task,
+      Object.assign(task, {
         agendaId: taskUpdates.agendaId || task.agendaId,
         content: taskUpdates.content || task.content,
         status: taskUpdates.status || task.status,
+        tags: taskUpdates.tags || task.tags,
         teamId: taskUpdates.teamId || task.teamId,
+        assigneeId: taskUpdates.assigneeId || task.assigneeId,
         assignee: taskUpdates.assigneeId
           ? this.db.teamMembers.find((teamMember) => teamMember.id === taskUpdates.assigneeId)
           : task.assignee,
         sortOrder: taskUpdates.sortOrder || task.sortOrder,
         userId: taskUpdates.userId || task.userId
       })
-      // keep the object the same
-      Object.assign(task, nextTask)
 
       const data = {
         __typename: 'UpdateTaskPayload',