+++ packages/client/mutations/CreateReflectionMutation.ts
@@ -22,6 +22,11 @@ graphql`
       meetingId
       sortOrder
       retroPhaseItemId
+      ...ReflectionGroupHeader_reflectionGroup @relay(mask: false)
+      ...ReflectionGroupVoting_reflectionGroup @relay(mask: false)
+      ...ReflectionGroup_reflectionGroup @relay(mask: false)
+      ...ReflectionGroupTitleEditor_reflectionGroup @relay(mask: false)
+      ...GroupingKanbanColumn_reflectionGroups @relay(mask: false)
       reflections {
         ...DraggableReflectionCard_reflection @relay(mask: false)
       }