+++ packages/client/utils/promiseAllObj.ts
@@ -1,4 +1,4 @@
-export default async (promiseObj) => {
+export default async (promiseObj: {[val: string]: Promise<any>}) => {
   const keys = Object.keys(promiseObj)
   const promises = Object.values(promiseObj)
   const values = await Promise.all(promises)