+++ packages/server/graphql/intranetSchema/intranetSchema.ts
@@ -15,6 +15,13 @@ import runScheduledJobs from './mutations/runScheduledJobs'
 import {GQLContext} from '../graphql'
 import loginSSO from './mutations/loginSSO'
 import draftEnterpriseInvoice from './mutations/draftEnterpriseInvoice'
+import githubAddAssignee from './mutations/githubAddAssignee'
+import stripeCreateInvoice from './mutations/stripeCreateInvoice'
+import stripeFailPayment from './mutations/stripeFailPayment'
+import stripeSucceedPayment from './mutations/stripeSucceedPayment'
+import stripeInvoiceFinalized from './mutations/stripeInvoiceFinalized'
+import stripeUpdateCreditCard from './mutations/stripeUpdateCreditCard'
+import stripeUpdateInvoiceItem from './mutations/stripeUpdateInvoiceItem'
 
 const query = new GraphQLObjectType<any, GQLContext, any>({
   name: 'Query',
@@ -36,10 +43,17 @@ const mutation = new GraphQLObjectType<any, GQLContext, any>({
     draftEnterpriseInvoice,
     endOldMeetings,
     flagOverLimit,
+    githubAddAssignee,
     loginSSO,
     runScheduledJobs,
     sendBatchNotificationEmails,
-    sendUpcomingInvoiceEmails
+    sendUpcomingInvoiceEmails,
+    stripeCreateInvoice,
+    stripeFailPayment,
+    stripeSucceedPayment,
+    stripeUpdateCreditCard,
+    stripeUpdateInvoiceItem,
+    stripeInvoiceFinalized,
   })
 })