+++ packages/server/graphql/mutations/updateReflectionContent.ts
@@ -14,6 +14,7 @@ import getReflectionEntities from './helpers/getReflectionEntities'
 import {GQLContext} from '../graphql'
 import getGroupSmartTitle from 'parabol-client/utils/autogroup/getGroupSmartTitle'
 import updateSmartGroupTitle from './helpers/updateReflectionLocation/updateSmartGroupTitle'
+import stringSimilarity from 'string-similarity'
 
 export default {
   type: UpdateReflectionContentPayload,
@@ -27,7 +28,7 @@ export default {
       description: 'A stringified draft-js document containing thoughts'
     }
   },
-  async resolve(_source, {reflectionId, content}, {authToken, dataLoader, socketId: mutatorId}: GQLContext) {
+  async resolve (_source, {reflectionId, content}, {authToken, dataLoader, socketId: mutatorId}: GQLContext) {
     const r = getRethink()
     const operationId = dataLoader.share()
     const now = new Date()
@@ -55,11 +56,11 @@ export default {
 
     // VALIDATION
     const normalizedContent = normalizeRawDraftJS(content)
-    const plaintextContent = extractTextFromDraftString(normalizedContent)
-    const isVeryDifferent = Math.abs(plaintextContent.length - reflection.plaintextContent.length) > 2
-    const entities = isVeryDifferent ? await getReflectionEntities(plaintextContent) : reflection.entities
 
     // RESOLUTION
+    const plaintextContent = extractTextFromDraftString(normalizedContent)
+    const isVeryDifferent =  stringSimilarity.compareTwoStrings(plaintextContent, reflection.plaintextContent) < 0.9
+    const entities = isVeryDifferent ? await getReflectionEntities(plaintextContent) : reflection.entities
     await r
       .table('RetroReflection')
       .get(reflectionId)
@@ -70,9 +71,9 @@ export default {
         updatedAt: now
       })
 
-    const reflectionsInGroup = await r.table('Reflection')
+    const reflectionsInGroup = await r.table('RetroReflection')
       .getAll(reflectionGroupId, {index: 'reflectionGroupId'})
-      .filter({isActive: true})
+      .filter({isActive: true}) as Reflection[]
 
     const newTitle = getGroupSmartTitle(reflectionsInGroup)
     await updateSmartGroupTitle(reflectionGroupId, newTitle)